Move user functionality into seperate model package
