+++ src/main/java/com/refactoringflow/refactoringflowbackend/controller/AssignmentController.java
@@ -2,7 +2,7 @@ package com.refactoringflow.refactoringflowbackend.controller;
 
 import com.refactoringflow.refactoringflowbackend.exception.ResourceNotFoundException;
 import com.refactoringflow.refactoringflowbackend.model.Assignment;
-import com.refactoringflow.refactoringflowbackend.model.Student;
+import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import com.refactoringflow.refactoringflowbackend.repository.AssignmentRepository;
 import com.refactoringflow.refactoringflowbackend.service.AssignmentService;
 import org.springframework.beans.factory.annotation.Autowired;
