+++ src/main/java/com/refactoringflow/refactoringflowbackend/controller/AuthenticationController.java
@@ -3,7 +3,10 @@ package com.refactoringflow.refactoringflowbackend.controller;
 import com.refactoringflow.refactoringflowbackend.config.SecurityConfig;
 import com.refactoringflow.refactoringflowbackend.error.exceptions.RefreshTokenException;
 import com.refactoringflow.refactoringflowbackend.exchanges.*;
-import com.refactoringflow.refactoringflowbackend.model.*;
+import com.refactoringflow.refactoringflowbackend.model.user.RefreshToken;
+import com.refactoringflow.refactoringflowbackend.model.user.Role;
+import com.refactoringflow.refactoringflowbackend.model.user.Student;
+import com.refactoringflow.refactoringflowbackend.model.user.User;
 import com.refactoringflow.refactoringflowbackend.service.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
