+++ src/main/java/com/refactoringflow/refactoringflowbackend/controller/CodeFileController.java
@@ -2,7 +2,7 @@ package com.refactoringflow.refactoringflowbackend.controller;
 
 import com.refactoringflow.refactoringflowbackend.ResponseMessage;
 import com.refactoringflow.refactoringflowbackend.model.CodeFile;
-import com.refactoringflow.refactoringflowbackend.model.Student;
+import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import com.refactoringflow.refactoringflowbackend.service.CodeFileService;
 import com.refactoringflow.refactoringflowbackend.service.StudentService;
 import org.springframework.beans.factory.annotation.Autowired;
