+++ src/main/java/com/refactoringflow/refactoringflowbackend/model/user/Role.java
@@ -1,4 +1,4 @@
-package com.refactoringflow.refactoringflowbackend.model;
+package com.refactoringflow.refactoringflowbackend.model.user;
 
 import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
