+++ src/main/java/com/refactoringflow/refactoringflowbackend/model/user/User.java
@@ -1,4 +1,4 @@
-package com.refactoringflow.refactoringflowbackend.model;
+package com.refactoringflow.refactoringflowbackend.model.user;
 
 import lombok.Getter;
 import lombok.NonNull;
