+++ src/main/java/com/refactoringflow/refactoringflowbackend/repository/AssignmentRepository.java
@@ -2,7 +2,7 @@ package com.refactoringflow.refactoringflowbackend.repository;
 
 
 import com.refactoringflow.refactoringflowbackend.model.Assignment;
-import com.refactoringflow.refactoringflowbackend.model.Student;
+import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
