+++ src/main/java/com/refactoringflow/refactoringflowbackend/service/AssignmentService.java
@@ -2,7 +2,7 @@ package com.refactoringflow.refactoringflowbackend.service;
 
 import com.refactoringflow.refactoringflowbackend.repository.AssignmentRepository;
 import com.refactoringflow.refactoringflowbackend.model.Assignment;
-import com.refactoringflow.refactoringflowbackend.model.Student;
+import com.refactoringflow.refactoringflowbackend.model.user.Student;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
