+++ src/main/java/com/refactoringflow/refactoringflowbackend/service/RoleService.java
@@ -1,6 +1,6 @@
 package com.refactoringflow.refactoringflowbackend.service;
 
-import com.refactoringflow.refactoringflowbackend.model.Role;
+import com.refactoringflow.refactoringflowbackend.model.user.Role;
 import com.refactoringflow.refactoringflowbackend.repository.RoleRepository;
 import org.springframework.stereotype.Service;
